Refactor API views
